Bug 1427892 — Move key commands into their own file.
