+++ Client/Frontend/Browser/TabTrayController.swift
@@ -278,6 +278,14 @@ class TabTrayController: UIViewController {
         return delegate
     }()
 
+    var numberOfColumns: Int {
+        return tabLayoutDelegate.numberOfColumns
+    }
+
+    var tabs: [Tab] {
+        return tabDataSource.tabs
+    }
+
     init(tabManager: TabManager, profile: Profile) {
         self.tabManager = tabManager
         self.profile = profile
@@ -416,7 +424,6 @@ class TabTrayController: UIViewController {
 
     func didClickAddTab() {
         openNewTab()
-        LeanPlumClient.shared.track(event: .openedNewTab, withParameters: ["Source": "Tab Tray" as AnyObject])
     }
 
     func didTapLearnMore() {
@@ -492,8 +499,13 @@ class TabTrayController: UIViewController {
             didTogglePrivateMode()
         }
     }
-    
-    fileprivate func openNewTab(_ request: URLRequest? = nil) {
+
+    func openNewTab() {
+        LeanPlumClient.shared.track(event: .openedNewTab, withParameters: ["Source": "Tab Tray" as AnyObject])
+        openNewTab(nil)
+    }
+
+    fileprivate func openNewTab(_ request: URLRequest?) {
         toolbar.isUserInteractionEnabled = false
 
         // We're only doing one update here, but using a batch update lets us delay selecting the tab
@@ -515,7 +527,7 @@ class TabTrayController: UIViewController {
         })
     }
 
-    fileprivate func closeTabsForCurrentTray() {
+    func closeTabsForCurrentTray() {
         tabManager.removeTabsWithUndoToast(tabsToDisplay)
         self.collectionView.reloadData()
     }
@@ -1003,74 +1015,6 @@ extension TabTrayController: UIAdaptivePresentationControllerDelegate, UIPopover
     // not as a full-screen modal, which is the default on compact device classes.
     func adaptivePresentationStyle(for controller: UIPresentationController, traitCollection: UITraitCollection) -> UIModalPresentationStyle {
         return .none
-    }
-}
-
-extension TabTrayController {
-    override var keyCommands: [UIKeyCommand]? {
-        let toggleText = privateMode ? Strings.SwitchToNonPBMKeyCodeTitle: Strings.SwitchToPBMKeyCodeTitle
-        return [
-            UIKeyCommand(input: "`", modifierFlags: .command, action: #selector(didTogglePrivateModeKeyCommand), discoverabilityTitle: toggleText),
-            UIKeyCommand(input: "w", modifierFlags: .command, action: #selector(didCloseTabKeyCommand)),
-            UIKeyCommand(input: "\u{8}", modifierFlags: [], action: #selector(didCloseTabKeyCommand), discoverabilityTitle: Strings.CloseTabFromTabTrayKeyCodeTitle),
-            UIKeyCommand(input: "w", modifierFlags: [.command, .shift], action: #selector(didCloseAllTabsKeyCommand), discoverabilityTitle: Strings.CloseAllTabsFromTabTrayKeyCodeTitle),
-            UIKeyCommand(input: "\r", modifierFlags: [], action: #selector(didEnterTabKeyCommand), discoverabilityTitle: Strings.OpenSelectedTabFromTabTrayKeyCodeTitle),
-            UIKeyCommand(input: "t", modifierFlags: .command, action: #selector(didOpenNewTabKeyCommand), discoverabilityTitle: Strings.OpenNewTabFromTabTrayKeyCodeTitle),
-            UIKeyCommand(input: UIKeyInputLeftArrow, modifierFlags: [], action: #selector(didChangeSelectedTabKeyCommand(sender:))),
-            UIKeyCommand(input: UIKeyInputRightArrow, modifierFlags: [], action: #selector(didChangeSelectedTabKeyCommand(sender:))),
-            UIKeyCommand(input: UIKeyInputDownArrow, modifierFlags: [], action: #selector(didChangeSelectedTabKeyCommand(sender:))),
-            UIKeyCommand(input: UIKeyInputUpArrow, modifierFlags: [], action: #selector(didChangeSelectedTabKeyCommand(sender:))),
-        ]
-    }
-
-    func didTogglePrivateModeKeyCommand() {
-        didTogglePrivateMode()
-    }
-
-    func didCloseTabKeyCommand() {
-        if let tab = tabManager.selectedTab {
-            tabManager.removeTab(tab)
-        }
-    }
-
-    func didCloseAllTabsKeyCommand() {
-        closeTabsForCurrentTray()
-    }
-
-    func didEnterTabKeyCommand() {
-        _ = self.navigationController?.popViewController(animated: true)
-    }
-
-    func didOpenNewTabKeyCommand() {
-        openNewTab()
-    }
-
-    func didChangeSelectedTabKeyCommand(sender: UIKeyCommand) {
-        let step: Int
-        switch sender.input {
-        case UIKeyInputLeftArrow:
-            step = -1
-        case UIKeyInputRightArrow:
-            step = 1
-        case UIKeyInputUpArrow:
-            step = -tabLayoutDelegate.numberOfColumns
-        case UIKeyInputDownArrow:
-            step = tabLayoutDelegate.numberOfColumns
-        default:
-            step = 0
-        }
-
-        let tabs = tabDataSource.tabs
-        let currentIndex: Int
-        if let selected = tabManager.selectedTab {
-            currentIndex = tabs.index(of: selected) ?? 0
-        } else {
-            currentIndex = 0
-        }
-
-        let nextIndex = max(0, min(currentIndex + step, tabs.count - 1))
-        let nextTab = tabs[nextIndex]
-        tabManager.selectTab(nextTab)
     }
 }
